ai-engine/requirements-pro.txt
# Requisitos para IA Profesional - Financial Copilot
# Compatible con Python 3.8-3.12

# Core ML (versiones compatibles)
scikit-learn>=1.2.0
pandas>=2.0.0
numpy>=1.23.0

# Advanced ML
xgboost>=1.7.0
lightgbm>=4.0.0

# Database
psycopg2-binary>=2.9.0

# API
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Utilities
python-dateutil>=2.8.0
joblib>=1.3.0
tqdm>=4.65.0
